FT2-112: Added autoloading and namespacing.
