+++ app/controllers/Loginwithgoogle.php
@@ -1,13 +1,16 @@
 <?php
-$model = new Model;
-$model->modelCall('user');
+namespace App\Controllers;
+
+use App\core\Database;
+use App\core\Controller;
+use App\models\User;
+
 class Loginwithgoogle
 {
     use Controller, User, Database;
 
     public function index()
     {
-
         if (isset($_SESSION["registered"]) && $_SESSION["registered"]) {
             header('Location: /public/welcome');
             exit;
