+++ app/controllers/Logout.php
@@ -1,5 +1,6 @@
 <?php
-require '../app/core/google-config.php';
+
+$google_client = new \Google_Client();
 
 $google_client->revokeToken();
 session_unset();
