+++ app/controllers/Welcome.php
@@ -1,8 +1,9 @@
 <?php
-$model = new Model;
-$model->modelCall('post');
-$model->modelCall('user');
+namespace App\Controllers;
 
+use App\models\User;
+use App\models\Post;
+use App\core\Controller;
 class Welcome
 {
     use Post, User, Controller;
