+++ app/core/Database.php
@@ -1,5 +1,5 @@
 <?php
-
+namespace App\core;
 trait Database
 {
     public $conn;
@@ -7,7 +7,7 @@ trait Database
     {
         $this->conn = mysqli_connect(DBHOST, DBUSER, DBPASS, DBNAME);
         if (!$this->conn) {
-            throw new Exception('Failed to connect to database: ' . mysqli_connect_error());
+            throw new \Exception('Failed to connect to database: ' . mysqli_connect_error());
         }
         return $this->conn;
     }
@@ -18,7 +18,7 @@ trait Database
         try {
             $result = mysqli_query($con, $sql);
             if ($result === false) {
-                throw new Exception("Query failed: " . mysqli_error($con));
+                throw new \Exception("Query failed: " . mysqli_error($con));
             }
             if (is_bool($result)) {
                 return true;
@@ -26,7 +26,7 @@ trait Database
                 return $result;
                 
             }
-        } catch (Exception $e) {
+        } catch (\Exception $e) {
             return $e->getMessage();
         } finally {
             mysqli_close($con);
