--- app/core/Model.php
@@ -1,9 +1,0 @@
-<?php
-class Model
-{
-    use Database;
-    public function modelCall($model)
-    {
-        require "../app/models/$model.php";
-    }
-}
