--- app/core/init.php
@@ -1,8 +1,0 @@
-<?php
-
-require 'config.php';
-require 'functions.php';
-require 'Database.php';
-require 'Model.php';
-require 'Controller.php';
-require 'App.php';
