+++ composer.json
@@ -2,5 +2,10 @@
     "require": {
         "phpmailer/phpmailer": "^6.9",
         "google/apiclient": "^2.15"
+    },
+    "autoload": {
+        "psr-4": {
+            "App\\": "app/" 
+        }
     }
 }
